When sending an update via the timeline, don't display the "we acknowledged you pressed enter"-equivalent message. That is, set the outgoing content object to not be displayed.
We still will error into the chat if it fails. This also adds an account preference (defaulted to on) to perform an update after sending a tweet. Our default refresh interval of 10 minutes gives us ≈65 tweets/hour below the maximum of 100 tweets/hour, so I don't think this will be at all detrimental.

This will currently cause problems if a tweet is sent during an update, which should be fixed in due time. See the TODO list in #11747 ("don't perform an update if we already have one in progress").