Add a new type of saved password—"special" passwords. Can be used for pretty much anything, but the only use of it now is the following.
Add a NickServ identification plugin, which stores the password in the keychain. When NickServ says the nick is identified, it tries the saved password. Failing its existence, or failing its verification, a new password is asked and stored if desired. Storing is defaulted to on.

This was only really tested on FreeNode; probably needs some modifications for other networks.