Throttle Growl messages. Fixes #12133.
We throttle messages every 0.5 seconds; every incoming message pushes the "clear" point back, preventing a huge rush of notifications at any one point in time.

Adds a "description for combined event" to all AIEventHandlers. Mostly they ignore the listObject and chat options for their descriptions, since the title is going to be the chat or listObject where appropriate in the Growl notification; other things might not necessarily have this requirement, so the option's there.